refactor: extract remaining read route registrars (#1732)
