+++ control_plane/http_routes/__init__.py
@@ -1,8 +1,27 @@
+from control_plane.http_routes.drivers import (
+    DriverReadRouteDependencies,
+    register_dokploy_target_inspect_read_routes,
+    register_driver_descriptor_read_routes,
+    register_operation_status_read_routes,
+    register_tracked_target_log_read_routes,
+)
+from control_plane.http_routes.every_code import (
+    register_every_code_feedback_read_routes,
+    register_every_code_notification_attempt_read_routes,
+    register_every_code_preview_gate_read_routes,
+    register_every_code_work_request_read_routes,
+)
 from control_plane.http_routes.ingress import register_ingress_read_routes
+from control_plane.http_routes.merge_train import register_merge_train_read_routes
 from control_plane.http_routes.operational_records import (
     register_deployment_promotion_read_routes,
     register_inventory_operation_read_routes,
     register_managed_secret_read_routes,
+)
+from control_plane.http_routes.preview import (
+    register_preview_notification_attempt_read_routes,
+    register_preview_readiness_read_routes,
+    register_preview_record_read_routes,
 )
 from control_plane.http_routes.products import (
     ProductReadRouteDependencies,
@@ -17,21 +36,42 @@ from control_plane.http_routes.products import (
 )
 from control_plane.http_routes.support import ReadRouteDependencies
 from control_plane.http_routes.topology import register_topology_read_routes
+from control_plane.http_routes.work_graph import (
+    WorkGraphReadRouteDependencies,
+    register_work_graph_issue_inbox_read_routes,
+    register_work_graph_snapshot_read_routes,
+)
 
 __all__ = (
+    "DriverReadRouteDependencies",
     "ProductReadRouteDependencies",
     "ReadRouteDependencies",
+    "WorkGraphReadRouteDependencies",
     "product_profile_context_cutover_contexts_allowed",
     "register_agent_context_read_routes",
     "register_deployment_promotion_read_routes",
+    "register_dokploy_target_inspect_read_routes",
+    "register_driver_descriptor_read_routes",
+    "register_every_code_feedback_read_routes",
+    "register_every_code_notification_attempt_read_routes",
+    "register_every_code_preview_gate_read_routes",
+    "register_every_code_work_request_read_routes",
     "register_ingress_read_routes",
     "register_inventory_operation_read_routes",
     "register_managed_secret_read_routes",
+    "register_merge_train_read_routes",
+    "register_operation_status_read_routes",
+    "register_preview_notification_attempt_read_routes",
+    "register_preview_readiness_read_routes",
+    "register_preview_record_read_routes",
     "register_product_config_status_read_routes",
     "register_product_context_audit_read_routes",
     "register_product_environment_read_routes",
     "register_product_profile_read_routes",
     "register_protected_artifact_read_routes",
     "register_topology_read_routes",
+    "register_tracked_target_log_read_routes",
+    "register_work_graph_issue_inbox_read_routes",
+    "register_work_graph_snapshot_read_routes",
     "require_product_profile_read_store",
 )
