+++ tests/test_http_read_route_registrars.py
@@ -22,11 +22,67 @@ class FastApiReadRouteRegistrarTests(unittest.TestCase):
     def test_extracted_routes_preserve_names_operation_ids_schemas_and_ownership(self) -> None:
         expected_routes = (
             (
+                "/v1/drivers/odoo/stable-bootstrap/operations/{operation_id}",
+                "read_odoo_stable_bootstrap_operation_status",
+                "read_odoo_stable_bootstrap_operation_status",
+                "OdooStableBootstrapOperationStatusResponse",
+                "control_plane.http_routes.drivers",
+            ),
+            (
+                "/v1/drivers/odoo/target-replacement/operations/{operation_id}",
+                "read_odoo_stable_target_replacement_operation_status",
+                "read_odoo_target_replacement_operation_status",
+                "OdooStableTargetReplacementOperationStatusResponse",
+                "control_plane.http_routes.drivers",
+            ),
+            (
                 "/v1/artifacts/protected",
                 "read_protected_artifacts",
                 "read_protected_artifacts",
                 "ProtectedArtifactsResponse",
                 "control_plane.http_routes.products",
+            ),
+            (
+                "/v1/drivers",
+                "read_driver_descriptors",
+                "read_driver_descriptors",
+                "DriverDescriptorsResponse",
+                "control_plane.http_routes.drivers",
+            ),
+            (
+                "/v1/drivers/{driver_id}",
+                "read_driver_descriptor",
+                "read_driver_descriptor",
+                "DriverDescriptorResponse",
+                "control_plane.http_routes.drivers",
+            ),
+            (
+                "/v1/contexts/{context}/driver-view",
+                "read_driver_context_view",
+                "read_driver_context_view",
+                "DriverContextViewResponse",
+                "control_plane.http_routes.drivers",
+            ),
+            (
+                "/v1/contexts/{context}/instances/{instance}/driver-view",
+                "read_driver_instance_view",
+                "read_driver_instance_view",
+                "DriverContextViewResponse",
+                "control_plane.http_routes.drivers",
+            ),
+            (
+                "/v1/dokploy-targets/inspect",
+                "read_dokploy_target_inspect",
+                "read_dokploy_target_inspect",
+                "DokployTargetInspectResponse",
+                "control_plane.http_routes.drivers",
+            ),
+            (
+                "/v1/contexts/{context}/instances/{instance}/logs",
+                "read_tracked_target_logs",
+                "read_tracked_target_logs",
+                "TrackedTargetLogsResponse",
+                "control_plane.http_routes.drivers",
             ),
             (
                 "/v1/deployments/{record_id}",
@@ -41,6 +97,27 @@ class FastApiReadRouteRegistrarTests(unittest.TestCase):
                 "read_promotion_record",
                 "PromotionRecordResponse",
                 "control_plane.http_routes.operational_records",
+            ),
+            (
+                "/v1/previews/readiness",
+                "read_preview_readiness",
+                "read_preview_readiness",
+                "PreviewReadinessResponse",
+                "control_plane.http_routes.preview",
+            ),
+            (
+                "/v1/previews/{preview_id}",
+                "read_preview_record",
+                "read_preview_record",
+                "PreviewRecordResponse",
+                "control_plane.http_routes.preview",
+            ),
+            (
+                "/v1/previews/{preview_id}/history",
+                "read_preview_history",
+                "read_preview_history",
+                "PreviewHistoryResponse",
+                "control_plane.http_routes.preview",
             ),
             (
                 "/v1/inventory/{context}/{instance}",
@@ -69,6 +146,90 @@ class FastApiReadRouteRegistrarTests(unittest.TestCase):
                 "read_agent_context",
                 "AgentContextResponse",
                 "control_plane.http_routes.products",
+            ),
+            (
+                "/v1/work-graph/snapshot",
+                "read_work_graph_snapshot",
+                "read_work_graph_snapshot",
+                "WorkGraphSnapshotResponse",
+                "control_plane.http_routes.work_graph",
+            ),
+            (
+                "/v1/work-graph/github/issues",
+                "read_work_graph_issue_inbox",
+                "read_work_graph_issue_inbox",
+                "WorkGraphIssueInboxResponse",
+                "control_plane.http_routes.work_graph",
+            ),
+            (
+                "/v1/work-graph/merge-train/admission",
+                "read_merge_train_admission",
+                "read_merge_train_admission",
+                "MergeTrainAdmissionResponse",
+                "control_plane.http_routes.merge_train",
+            ),
+            (
+                "/v1/work-graph/merge-train/controller/status",
+                "read_merge_train_controller_status",
+                "read_merge_train_controller_status",
+                "MergeTrainControllerStatusResponse",
+                "control_plane.http_routes.merge_train",
+            ),
+            (
+                "/v1/work-graph/merge-train/policy-targets",
+                "read_merge_train_policy_targets",
+                "read_merge_train_policy_targets",
+                "MergeTrainPolicyTargetsResponse",
+                "control_plane.http_routes.merge_train",
+            ),
+            (
+                "/v1/every-code/summary",
+                "read_every_code_summary",
+                "read_every_code_summary",
+                "EveryCodeSummaryResponse",
+                "control_plane.http_routes.every_code",
+            ),
+            (
+                "/v1/every-code/work-requests",
+                "list_every_code_work_requests",
+                "list_every_code_work_requests",
+                "EveryCodeWorkRequestRecordsResponse",
+                "control_plane.http_routes.every_code",
+            ),
+            (
+                "/v1/every-code/work-requests/{request_id}",
+                "read_every_code_work_request",
+                "read_every_code_work_request",
+                "EveryCodeWorkRequestRecordResponse",
+                "control_plane.http_routes.every_code",
+            ),
+            (
+                "/v1/every-code/pr-feedback",
+                "list_every_code_pr_feedback",
+                "list_every_code_pr_feedback",
+                "EveryCodePrFeedbackRecordsResponse",
+                "control_plane.http_routes.every_code",
+            ),
+            (
+                "/v1/every-code/preview-gates",
+                "list_every_code_preview_gates",
+                "list_every_code_preview_gates",
+                "EveryCodePreviewGateRecordsResponse",
+                "control_plane.http_routes.every_code",
+            ),
+            (
+                "/v1/every-code/notification-attempts",
+                "list_every_code_notification_attempts",
+                "list_every_code_notification_attempts",
+                "EveryCodeNotificationAttemptRecordsResponse",
+                "control_plane.http_routes.every_code",
+            ),
+            (
+                "/v1/previews/pr-feedback/notification-attempts",
+                "list_preview_pr_feedback_notification_attempts",
+                "list_preview_pr_feedback_notification_attempts",
+                "PreviewPrFeedbackNotificationAttemptRecordsResponse",
+                "control_plane.http_routes.preview",
             ),
             (
                 "/v1/products",
